# machinev2/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Law as Code
  version: 0.0.1
servers:
  - url: https://api.server.test/v1
...
paths:
  /evaluate:
    post:
      tags:
        - Law
      description: Evaluate the law
      operationId: evaluate
      requestBody:
        $ref: "#/components/requestBodies/EvaluateRequest"
      responses:
        "201":
          $ref: "#/components/responses/EvaluateResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
components:
  schemas:
...
    Error:
      description: The error that occured while processing this request.
      type: object
      properties:
        message:
          type: string
          example: foo went wrong
      required:
        - message
    Evaluate:
      description: Evaluate.
      type: object
      properties:
        service:
          type: string
          description: Specify the service that needs to be executed
        law:
          type: string
          description: Specify the law that needs to be executed
        parameters:
          type: object
          additionalProperties: true
        date:
          type: string
          format: date
          description: Can be used to overwrite the date used by the service
          example: "2025-01-31"
        input:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        output:
          type: string
          description: specify a requested output value
        approved:
          type: boolean
          description: only use approved claims, default to true
      required:
        - service
        - law
      example:
        service: "TOESLAGEN"
        law: "zorgtoeslagwet"
    EvaluateResponseSchema:
      description: Evaluate response
      type: object
      properties:
        output:
          type: object
          additionalProperties: true
        input:
          type: object
          additionalProperties: true
        requirementsMet:
          type: boolean
          description: Will be true when all requirements where met
        rulespecId:
          type: string
          format: uuid
          description: Identifier of the rulespec
        missingRequired:
          type: boolean
          description: Will be true when a required value is missing
        path:
          $ref: "#/components/schemas/PathNode"
      required:
        - output
        - input
        - requirementsMet
        - rulespecId
        - missingRequired
        - path
    PathNode:
      description: path node
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        result: {}
        resolveType:
          type: string
        required:
          type: boolean
        details:
          type: object
          additionalProperties: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/PathNode"
      required:
        - type
        - name
...
  requestBodies:
    EvaluateRequest:
      description: An evaluation request.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: "#/components/schemas/Evaluate"
            required:
              - data
...
  responses:
...
    BadRequestErrorResponse:
      description: Invalid input supplied.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/Error"
                example:
                  - message: foo went wrong
            required:
              - errors
            example:
              errors:
                - message: foo went wrong
...
    InternalServerErrorResponse:
      description: Internal server error.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/Error"
                example:
                  - message: foo went wrong
            required:
              - errors
            example:
              errors:
                - message: foo went wrong
    EvaluateResponse:
      description: Evaluate response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: "#/components/schemas/EvaluateResponseSchema"
            required:
              - data
...